src/jobstat-hub/requirements.txt
rejson==0.5.4
kafka-python==2.0.2